Copy modules from bdp-iceberg
Copy metacat and view module from bdp-iceberg.

NETFLIX-BUILD: Making bdp-iceberg dependent on apache-iceberg

In order to make this move we had to implement the Catalog API and
create a iceberg-spark module which has netflix specific spark source
that is built using spark-2.1.1 where as the open source version is
using spark-2.4.4.