Common View Support Library
This commit adds the library to support common views across Presto and Spark.
The library provides an interface to create metastore managed views and hadoop views (for testing).
The interface allows:
+ Create of a view
+ Replace of a view
+ Load of a view
+ Drop of a view
+ Add/Delete/Edit column comments
The metadata of the view is stored in JSON format on S3. Each create or replace of the view results
in creation of a new view version metadata file. The current view version metdata file shows last
few versions. This number is controlled by the table property "version.history.num-entries".

The High Level Design Doc can be found here:
https://docs.google.com/document/d/1-LlYx8Q1QH-GFZ4hlxp3__x5lDF-cs3QrQUyBPz4AJ8/edit?usp=sharing